requirements.txt
alembic==1.8.1
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
Click==8.0.1
et-xmlfile==1.0.1
Flask==2.2.2
Flask-Cors==3.0.10
gevent==22.10.2
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==2.0.1
gunicorn==20.1.0
idna==2.9
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.2
Mako==1.1.0
MarkupSafe==2.1.1
numpy==1.26.0
openpyxl==3.1.2
pandas==1.4.0
protobuf==3.11.3
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
python-dateutil==2.8.1
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.8
Werkzeug==2.2.2
zope.event==4.4
zope.interface==5.1.0

